Add html elements and logic
RPS can now be played on the browser. The score resets once it reaches 5 and announces the victor of the game. Functions were added and simplified to follow function principle of one task only.

I plan to make a button that would restart the entire game once the victor is announced. Also, it lets the player see who scored 5 aside from the victor announcement.